moved util functions into separate file (utils.tsx)
